+++ src/components/Footer.js
@@ -151,6 +151,9 @@ const Wrapper = styled.div`
   align-items: center;
   justify-content: space-around;
   box-shadow: 0px -2px 4px rgb(0 0 0 / 15%);
+  @media (max-width: 651px) {
+    display: none;
+  }
 `;
 
 const LogoContainer = styled.div`
